+++ python/dartpy/collision/module.cpp
@@ -4,6 +4,7 @@
 #include "collision/collision_group.hpp"
 #include "collision/collision_option.hpp"
 #include "collision/collision_result.hpp"
+#include "collision/continuous_collision.hpp"
 #include "collision/raycast.hpp"
 
 namespace dart::python_nb {
@@ -12,8 +13,9 @@ void defCollisionModule(nanobind::module_& m)
 {
   defCollisionOption(m);
   defCollisionResult(m);
-  defRaycast(m);
   defCollisionDetector(m);
+  defRaycast(m);
+  defContinuousCollision(m);
   defCollisionGroup(m);
 }